Wrote specs for all methods.

+++ src/emeralds/yaml_processor.cr
@@ -1,5 +1,9 @@
 require "yaml"
 
+##
+# @class: YamlProcessor
+# @brief Helper functions for reading from the em.yml file
+##
 class Emeralds::YamlProcessor
     ##
     # @message: get_dependencies
@@ -7,13 +11,16 @@ class Emeralds::YamlProcessor
     # @return -> The list of dependencies
     ##
     def get_dependencies : Array(String)
-        yaml = File.open("em.yml") { |f| YAML.parse f; }.as_h;
-
-        if !yaml["dependencies"]
-            [] of String;
+        if File.exists?("em.yml")
+            yaml = File.open("em.yml") { |f| YAML.parse f; }.as_h;
+            if !yaml["dependencies"]
+                [] of String;
+            else
+                # Remove brackets and split at commas
+                yaml["dependencies"].to_s.lstrip("{").rstrip("}").split(", ");
+            end
         else
-            # Remove brackets and split at commas
-            yaml["dependencies"].to_s.lstrip("{").rstrip("}").split(", ");
+            [] of String;
         end
     end
 
@@ -32,6 +39,11 @@ class Emeralds::YamlProcessor
         end
     end
 
+    ##
+    # @message: get_lines_of_code
+    # @brief Read all source files and count the lines of codes
+    # @return -> The total number of files nad lines of code
+    ##
     def get_lines_of_code : Array(Int32)
         num = 0;
         loc = 0;
